# app/auth/api.py
try:
    from fastapi import APIRouter, Request, HTTPException
except ImportError:
    APIRouter = Request = HTTPException = None
from app.auth.jwt import create_jwt_token
import uuid

router = APIRouter()

@router.post('/auth/token')
async def get_token(request: Request):
    data = await request.json()
    api_key = data.get('api_key')
    tenant_id = data.get('tenant_id')
    user_id = data.get('user_id', str(uuid.uuid4()))
    # TODO: Validate api_key and tenant_id
    if not api_key or not tenant_id:
        raise HTTPException(status_code=400, detail='Missing api_key or tenant_id')
    # For demo, accept any api_key
    token = create_jwt_token(tenant_id, user_id)
    return {"access_token": token, "token_type": "bearer"}