# app/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Unified Revenue Context Engine API
  version: 1.0.0
  description: |
    API for multi-tenant, compliance-native, production-grade B2B revenue orchestration platform.
servers:
  - url: https://api.yourdomain.com/api
paths:
  /auth/token:
    post:
      summary: Exchange API key for JWT
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                api_key:
                  type: string
                tenant_id:
                  type: string
      responses:
        '200':
          description: JWT token
...